requirements.txt
keras==3.5.0
matplotlib==3.7.2
matplotlib-inline==0.1.6
numpy==1.23.5
opencv-python==4.10.0.84
pandas==2.0.3
pyserial==3.5
tensorflow==2.17.0
